+++ utilities/pycvm/pycvm/elevation_cross_section.py
@@ -200,7 +200,7 @@ class ElevationCrossSection:
 
         scale_gate = None
         if 'gate' in self.meta :
-           scale_gate = int(self.meta['gate'])
+           scale_gate = float(self.meta['gate'])
 
         if color_scale == "b" and scale_gate is None:
            scale_gate=2.5
