+++ utilities/pycvm/pycvm/horizontal_slice.py
@@ -185,7 +185,7 @@ class HorizontalSlice:
            color_scale = self.meta['color']
 
         if 'gate' in self.meta :
-           scale_gate = int(self.meta['gate'])
+           scale_gate = float(self.meta['gate'])
         
         if color_scale == "b" and scale_gate is None:
            scale_gate=2.5
